requirements.txt
faster-whisper==0.7.1
potassium
torch==1.13.0
transformers==4.25.1
accelerate==0.15.0
ffmpeg-python==0.2.0
pandas==1.5.0
torchtext==0.14.0
torchvision==0.14.0
sacremoses==0.0.53
sentencepiece==0.1.97
tokenizers==0.13.2
tqdm==4.64.1
EasyNMT==2.0.2
nltk==3.8.1
transformers==4.25.1
pysrt==1.1.2
psutil==5.9.2
requests==2.28.2
pyannote.audio==2.1.1
python-magic==0.4.27